continuing to move files around and moving tuples to their own files
