+++ app/replayer/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 
 add_subdirectory(lib)
-add_subdirectory(app)
+#add_subdirectory(app)
 if(WMTK_APP_UNIT_TESTS)
     add_subdirectory(tests)
 endif()
