+++ src/wmtk/serialization/TriMeshOperationLogger.cpp
